pyproject.toml
[project]
name = "self-healing-crypto-trader"
version = "1.0.0"
description = "An autonomous AI crypto trader that heals itself by having Claude analyze its own trading logs and improve strategies over time"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "websocket-client>=1.7.0",
    "pydantic>=2.0.0",
]

[project.scripts]
start = "src.main:main"
